Move Lttng provider into CMake tree.
