+++ src/pal/src/CMakeLists.txt
@@ -409,5 +409,10 @@ endif(CMAKE_SYSTEM_NAME STREQUAL NetBSD)
 
 add_subdirectory(examples)
 
+if(FEATURE_EVENT_TRACE)
+  add_subdirectory(eventprovider)
+endif(FEATURE_EVENT_TRACE)
+
+
 # Install the static PAL library for VS
 _install (TARGETS coreclrpal DESTINATION lib)
