+++ src/pal/src/eventprovider/lttngprovider/CMakeLists.txt
@@ -3,7 +3,7 @@ cmake_minimum_required(VERSION 2.8.12)
 include(FindPythonInterp)
 
 execute_process(
-  COMMAND ${PYTHON_EXECUTABLE} ${CMAKE_CURRENT_SOURCE_DIR}/../../../scripts/genLttngProvider.py --man ${EVENT_MANIFEST} --intermediate ${CMAKE_CURRENT_BINARY_DIR}
+  COMMAND ${PYTHON_EXECUTABLE} ${CMAKE_CURRENT_SOURCE_DIR}/../../../../scripts/genLttngProvider.py --man ${EVENT_MANIFEST} --intermediate ${CMAKE_CURRENT_BINARY_DIR}
   RESULT_VARIABLE GEN_LTTNG_RESULT
   OUTPUT_VARIABLE LTTNG_PROVIDER_OUTPUT
   ERROR_VARIABLE GEN_LTTNG_ERRORS
@@ -31,14 +31,13 @@ set(CMAKE_INCLUDE_CURRENT_DIR ON)
 
 add_definitions(-DPAL_STDCPP_COMPAT=1)
 include_directories(${COREPAL_SOURCE_DIR}/inc/rt)
-include_directories(lttng)
+include_directories(${CMAKE_CURRENT_BINARY_DIR}/lttng)
 
 add_library(eventprovider
     STATIC
     ${LTTNG_PROVIDER_SOURCES}
+    eventproviderhelpers.cpp
 )
-set_target_properties(eventprovider PROPERTIES LINKER_LANGUAGE CXX)
-
 add_subdirectory(tracepointprovider)
 
 # Install the static eventprovider library
